QCamera2: HAL: Modify instant AEC param to session based.
Issue:
Instance AEC is triggered for every restart of
preview in one camera open session.
Due to this fast AEC convergence algorithm is
performing during preview restart and AEC is not
settled properly.

Fix:
Modify the instant AEC param to session based param.
Dont update the state of instant AEC to app if HAL
is internally setting it.

CRs-fixed: 1023956
Change-Id: I7bc44ec55a5b6235c4ec128a74722750cf5bbb0b