QCamera2: Fix for LSS snapshot when flash is ON
Issue:
In flash on case HAL gives preference to LLS snapshot
instead of flash and as a result flash is not fired.

Fix:
Made changes to ensure flash is not turned off if LLS
is detected. Also made changes to ensure when flash is
auto if LLS is detected LLS snapshot is triggered and
not flash.

CRs-fixed: 1048636
Change-Id: Ide0f0ef1bb5623607aa2447852636a23f60a8650